Player permissions

+++ scripts/ui/questtracker.js
@@ -1,4 +1,4 @@
-import { constants } from "../helpers/global.js";
+import { constants, getSocket } from "../helpers/global.js";
 import { QuestDatabase } from "../data/database.js";
 import { QuestEditor } from "./questeditor.js";
 import { objectiveState } from "../helpers/global.js";
@@ -92,6 +92,9 @@ export class QuestTracker extends Application {
             collapsed: this.collapsed,
             quests: QuestDatabase.quests,
             isGM: game.user.isGM,
+            canEdit: game.user.isGM || Settings.get(Settings.NAMES.PLAYER_EDIT),
+            canCreate:
+                game.user.isGM || Settings.get(Settings.NAMES.PLAYER_CREATE),
             activeQuests: this.activeQuests,
             offset: `${Settings.get(Settings.NAMES.TRACKER_OFFSET) / 16}rem`,
             docked: Settings.get(Settings.NAMES.TRACKER_DOCKED),
@@ -146,6 +149,7 @@ export class QuestTracker extends Application {
         $html.find(".minimize").on("click", (evt) => {
             evt.stopPropagation();
             this.collapse();
+            console.log(result);
         });
 
         // Expand/Collapse quest bodies.
@@ -190,8 +194,6 @@ export class QuestTracker extends Application {
             this.render();
         });
 
-        // All of the listeners beyond this point are for the GM only.
-        if (!game.user.isGM) return;
         // Progress the state of an objective.
         $html.find(".simpler-quest-objective").on("click", (evt) => {
             evt.stopPropagation();
@@ -207,43 +209,29 @@ export class QuestTracker extends Application {
             // Update the objective state
             const questId =
                 evt.target.closest("[data-quest-id]").dataset.questId;
-            Quest.updateObjective(questId, objId, "state");
+
+            if (game.user.isGM) Quest.updateObjective(questId, objId, "state");
+            else {
+                getSocket().emit("UpdateObjective", {
+                    questId: questId,
+                    objId: objId,
+                    key: "state",
+                });
+            }
         });
 
-        // Toggle the Secret status of an objective.
-        $html.find(".simpler-quest-objective").on("contextmenu", (evt) => {
+        // All of the listeners beyond this point require permission.
+        if (!game.user.isGM && !Settings.get(Settings.NAMES.PLAYER_EDIT))
+            return;
+
+        // Create a new quest.
+        $html.find(".header > .new-quest").on("click", (evt) => {
             evt.stopPropagation();
-            evt.preventDefault();
 
-            // Get the object ID.
-            const objId = getObjectiveId(evt.target);
-            if (!objId) return;
-
-            // Update the objective secret
-            const questId =
-                evt.target.closest("[data-quest-id]").dataset.questId;
-            Quest.updateObjective(questId, objId, "secret");
+            // Load a blank QuestEditor and render.
+            let qe = new QuestEditor();
+            qe.render(true, { focus: true });
         });
-
-        // Toggle the visibility of quests.
-        $html
-            .find(".simpler-tracked-quest > .header > .vis-toggle")
-            .on("click", (evt) => {
-                evt.stopPropagation();
-
-                // Get the quest data.
-                const data = getQuestData(evt.target);
-
-                // If the quest data is valid
-                if (data.quest) {
-                    // Set the quest visibility and save.
-                    QuestDatabase.update({
-                        ...data.quest,
-                        visible: !data.quest.visible,
-                    });
-                    QuestDatabase.save();
-                }
-            });
 
         // Edit a quest's data.
         $html
@@ -259,15 +247,6 @@ export class QuestTracker extends Application {
                 let qe = new QuestEditor({ questId: questId });
                 qe.render(true, { focus: true });
             });
-
-        // Create a new quest.
-        $html.find(".header > .new-quest").on("click", (evt) => {
-            evt.stopPropagation();
-
-            // Load a blank QuestEditor and render.
-            let qe = new QuestEditor();
-            qe.render(true, { focus: true });
-        });
 
         // Delete a quest.
         $html
@@ -292,9 +271,50 @@ export class QuestTracker extends Application {
 
                 // If the user confirmed, delete the quest.
                 if (confirmed) {
-                    QuestDatabase.removeQuest(data.questId);
+                    if (game.user.isGM) QuestDatabase.removeQuest(data.questId);
+                    else {
+                        getSocket().emit("DeleteQuest", { questId: questId });
+                    }
                 }
             });
+
+        // Listeners beyond this point are GM Only.
+        if (!game.user.isGM) return;
+
+        // Toggle the visibility of quests.
+        $html
+            .find(".simpler-tracked-quest > .header > .vis-toggle")
+            .on("click", (evt) => {
+                evt.stopPropagation();
+
+                // Get the quest data.
+                const data = getQuestData(evt.target);
+
+                // If the quest data is valid
+                if (data.quest) {
+                    // Set the quest visibility and save.
+                    QuestDatabase.update({
+                        ...data.quest,
+                        visible: !data.quest.visible,
+                    });
+                    QuestDatabase.save();
+                }
+            });
+
+        // Toggle the Secret status of an objective.
+        $html.find(".simpler-quest-objective").on("contextmenu", (evt) => {
+            evt.stopPropagation();
+            evt.preventDefault();
+
+            // Get the object ID.
+            const objId = getObjectiveId(evt.target);
+            if (!objId) return;
+
+            // Update the objective secret
+            const questId =
+                evt.target.closest("[data-quest-id]").dataset.questId;
+            Quest.updateObjective(questId, objId, "secret");
+        });
     }
 
     refresh() {
